Move chart functions back to scripts. Add helper function to refactor renderWeekly function. Get charts working but still need better display
